refactor static app so props arent passed through App component, add CartNavItem and ProductTable components
