+++ src/index.js
@@ -4,19 +4,9 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
-const PRODUCTS = [
-  { name: 'apple', price: '$0.79', stocked: true },
-  { name: 'banana', price: '$0.50', stocked: true },
-  { name: 'peach', price: '$0.85', stocked: false },
-  { name: 'watermelon', price: '$1.50', stocked: true },
-  { name: 'mango', price: '$2.00', stocked: false }
-];
-
-console.log('PRODUCTS obj in index: ', PRODUCTS);
-
 ReactDOM.render(
   <React.StrictMode>
-    <App products={PRODUCTS}/>
+    <App/>
   </React.StrictMode>,
   document.getElementById('root')
 );
